requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==5.5
decorator==4.4.2
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
Flask==3.0.3
gunicorn==23.0.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
moviepy==1.0.3
mutagen==1.47.0
numpy==2.0.1
packaging==24.1
pillow==10.4.0
proglog==0.1.10
pycryptodomex==3.20.0
requests==2.32.3
toml==0.10.2
tqdm==4.66.4
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
yt-dlp==2025.3.31
